feat: Lightdash dashboards + multi-store Shopify + geography analytics
dbt:
- Add geography fields to fct_orders and fct_cin7_sales (uk_region,
  shipping_city/postcode/country via LATERAL FLATTEN on ADDRESSES array)
- New gold models: fct_product_sales, fct_product_pairs,
  fct_product_repeat_rate, fct_cin7_sales, dim_cin7_customers
- Cast is_first_order to integer in fct_orders and fct_cin7_sales so
  SUM() works in Snowflake (cannot SUM a boolean)
- Add Lightdash meta (metrics, labels) to gold _schema.yml for all new
  and existing models; add geography columns to fct_orders and
  fct_cin7_sales schema entries
- Add silver _schema.yml tests for new staging models
- Add bronze_prospect_crm source to _sources.yml

Lightdash:
- lightdash/build_dashboards.py — script to build all 6 dashboards via
  REST API (Group Overview, Returning Customers, Cross-Brand Customers,
  UK Regional Performance, Product Performance, Product Repeat Rate)

Airbyte:
- Document Revitalash and Geske Shopify connections (Bronze schemas
  exist; Revitalash full refresh in progress, Geske pending)
- Add cleanup script for potential BRONZE_SHOPIFY_ISCLINICAL
  contamination from misconfigured Revitalash namespace
- Update README with connection inventory and provisioning notes

CLAUDE.md:
- Document all four Shopify store connections and their sync status
- Add is_first_order integer cast convention
- Add product Gold models to architecture overview

+++ dbt/models/gold/fct_cin7_sales.sql
@@ -91,5 +91,5 @@ with_order_num as (
 
 select
     *,
-    customer_order_number = 1 as is_first_order
+    (customer_order_number = 1)::integer as is_first_order
 from with_order_num
